changed the type for duration to Long

+++ domain/src/main/java/com/chase/kudzie/chasemusic/domain/model/Song.kt
@@ -5,7 +5,7 @@ data class Song(
     val albumName: String,
     val artistId: Long,
     val artistName: String,
-    val duration: Int,
+    val duration: Long,
     val title: String,
     val trackNumber: Int
 )
